Smoothing using several methods

Smoothing using several methods

Usage

smoothing(
  x,
  methods = c("henderson", "henderson_localic", "henderson_robust",
    "henderson_robust_localic", "clf_cn", "clf_alf"),
  endpoints = "Musgrave",
  length = NULL,
  icr = NULL,
  asymmetric_var = FALSE,
  degree = 3,
  ao = NULL,
  ao_tc = NULL,
  ls = NULL,
  ...
)

Arguments

x

input time-series.

methods

list of methods to use.

endpoints

Method used for the asymmetric filter. By default the Musgrave method is used

length

the length of the

icr

I/C ratio used for the asymmetric filter.

asymmetric_var

when local_icr = TRUE, if asymmetric_var = TRUE then the variance is estimated for each asymmetric filters instead of using the variance associated to the symmetric estimates.

degree

if local_icr = TRUE, degree of polynomial used to estimate the local bias parameter.

ao

Dates of the Additive Outliers (AO) which effects are associated to the irregular component.

ao_tc

Dates of the Additive Outliers (AO) which effects are associated to the trend-cycle component.

ls

Dates of the Level Shifts (LS) which effects are associated to the trend-cycle component.

...

other unused parameters.