X-11 Selection of Trend-Cycle Filter

Perform X-11 selection of the length of Henderson (x11_trend_selection()) and compute the associated I/C ratio used to build Musgrave fuilters (find_icr()).

Usage

x11_trend_selection(x)

find_icr(length, freq = 12)

Arguments

x

a "ts" object.

length

length of the filter.

freq

frequency of the time series used to compute the I/C ratio.

Details

The following procedure is used in X-11 to select the length of the trend filter:

Computes the I/C ratio, \(icr\) with an Henderson filter of length the frequency plus 1.

The length depends on the value or \(icr\):

if \(icr < 1\) then the selected length is 9 for monthly data and 5 otherwise;

if \(1 \leq icr < 3.5\) then the selected length is \(freq + 1\) where \(freq\) is the frequency of data (12 for monthly data, 4 for quarterly data...).

if \(icr \geq 3.5\) then the selected length is 23 for monthly data and 7 otherwise.

The value of \(icr\) is then fixed to build Musgrave filters (find_icr()) :

for quarterly data, if the length is 5 then \(icr = 0.001\), otherwide \(icr = 4.5\);

if the length if less or equal to 9 then \(icr = 1\);

else if the length if less or equal to 13 then \(icr = 3.5\);

else \(icr = 4.5\).